config-manager/requirements.txt
fastapi ==0.110.0
uvicorn[standard] >=0.12.0,<0.23.0
SQLAlchemy ==2.0.29
python-dotenv ==1.0.1
pyctuator ==1.2.0
langchain ==0.2.1
langchainhub ==0.1.17
langchain-community ==0.2.1
langchain-google-vertexai ==1.0.4
faiss-cpu ==1.8.0
tiktoken ==0.7.0
google-cloud-aiplatform ==1.53.0
google-cloud-bigquery ==3.23.1
PyMuPDF ==1.24.7
python-docx ==1.1.2
pytest ==8.2.2
httpx ==0.23.1
jwt ==1.3.1
python_multipart ==0.0.9
passlib ==1.7.4
python-jose ==3.3.0
bcrypt ==4.2.0
psycopg2
email-validator